FreshBytes/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg
psycopg2-binary
Pillow
django-redis
environs
django-extensions
drf-spectacular
